+++ src/test/java/nextstep/subway/common/Common.java
@@ -1,0 +1,25 @@
+package nextstep.subway.common;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.springframework.http.MediaType;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class Common {
+
+    public static ExtractableResponse<Response> 지하철역을_생성한다(String stationName) {
+        Map<String, String> params = new HashMap<>();
+        params.put("name", stationName);
+
+        return RestAssured.given()
+                .body(params).log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().post("/stations")
+                .then()
+                .log().all()
+                .extract();
+    }
+}
